Extract PR lookup into shared package
